ex) #이슈번호, #이슈번호

## 📝 요약(Summary)

변경 사항 및 관련 이슈에 대해 간단하게 작성해주세요. 어떻게보다 무엇을 왜 수정했는지 설명해주세요.

## 🛠️ PR 유형

어떤 변경 사항이 있나요?

- [ ] 새로운 기능 추가
- [ ] 버그 수정
- [ ] CSS 등 사용자 UI 디자인 변경
- [ ] 코드에 영향을 주지 않는 변경사항(오타 수정, 탭 사이즈 변경, 변수명 변경)
- [ ] 코드 리팩토링
- [ ] 주석

+++ Backend/IsCream/src/main/java/com/ssafy/iscream/s3/service/S3Service.java
@@ -50,30 +50,34 @@ public class S3Service {
     }
 
     // 단일 파일 업로드
-    public String uploadImage(MultipartFile multipartFile) throws IOException {
+    public String uploadImage(MultipartFile multipartFile) {
         String fileKey = imageDir + FileUtil.createFileName(multipartFile);
 
-        // 임시 파일 생성
-        File tempFile = File.createTempFile("upload-", ".tmp");
-        multipartFile.transferTo(tempFile);
+        File tempFile = null;
 
         try {
+            // 임시 파일 생성
+            tempFile = File.createTempFile("upload-", ".tmp");
+            multipartFile.transferTo(tempFile);
             // S3에 업로드
             s3Client.putObject(
                     PutObjectRequest.builder().bucket(bucket).key(fileKey).build(),
                     RequestBody.fromFile(tempFile)
             );
-        } catch (SdkClientException e) {
+        } catch (IOException | SdkClientException e) {
             throw new S3UploadException(FILE_UPLOAD_FAILED);
         } finally {
-            tempFile.delete(); // 임시 파일 삭제
+            // 임시 파일 삭제
+            if (tempFile != null) {
+                tempFile.delete();
+            }
         }
 
         return baseUrl + fileKey; // 만들어진 url 반환
     }
 
     // 여러 파일 업로드
-    public List<String> uploadImage(List<MultipartFile> files) throws IOException {
+    public List<String> uploadImage(List<MultipartFile> files) {
         List<String> result = new ArrayList<>();
 
         for (MultipartFile file : files) {
